pyproject.toml
[tool.poetry]
name = "tradingbot-scheduler"
version = "0.1.0"
description = ""
authors = ["Justin Guese <guese.justin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
SQLAlchemy = "^2.0.23"
requests = "^2.31.0"
psycopg2-binary = "^2.9.9"
croniter = "^2.0.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
